requirements.txt
pillow
wand
textract
pytesseract
pyocr
Tesseract
numpy
opencv-python
imutils
matplotlib
argparse